+++ test/DefifaUSDC.t.sol
@@ -176,7 +176,7 @@ contract DefifaUSDCTest is JBTest, TestBaseWorkflow {
         uint8 n,
         uint104 tierPrice,
         uint256 minParticipation,
-        uint32 scorecardTimeout
+        uint256 scorecardTimeout
     )
         internal
         returns (DefifaLaunchProjectData memory)
