+++ test/TestAuditGaps.sol
@@ -173,7 +173,7 @@ contract TestAuditGapsERC20Games is JBTest, TestBaseWorkflow {
         uint8 n,
         uint104 tierPrice,
         uint256 minParticipation,
-        uint32 scorecardTimeout
+        uint256 scorecardTimeout
     )
         internal
         returns (DefifaLaunchProjectData memory)
